### 2025-06-20 #8439 @editorialbot
[PRE REVIEW]: ETLForge: A unified framework for synthetic test-data generation and ETL validation
**Submitting author:** @kkartas (<a href="http://orcid.org/0009-0001-6477-4676">Kyriakos Kartas</a>)
**Repository:** https://github.com/kkartas/ETLForge
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
