### 2025-06-21 d563c36
Add status badge to README.md for JOSS paper and update publisher details in paper.bib for consistency and accuracy.
