+++ paper/paper.bib
@@ -17,8 +17,7 @@
   title={Exploratory Data Mining and Data Cleaning},
   author={Dasu, Tamraparni and Johnson, Theodore},
   year={2003},
-  publisher={John Wiley & Sons},
-  doi={10.1002/0471725527}
+  publisher={John Wiley \& Sons}
 }
 
 @book{Loshin2010,
@@ -100,7 +99,8 @@
   number={12},
   pages={1781--1794},
   year={2018},
-  publisher={VLDB Endowment}
+  publisher={VLDB Endowment},
+  doi={10.14778/3229863.3229867}
 }
 
 @misc{GreatExpectations2023,
