### 2025-06-21 #8439 @crvernon
:wave: @kkartas - We are currently managing a large queue of submissions and will get you a topic editor assigned as soon as one becomes available.

In the meantime, your README doesn't state what ETL is.  I know this is in your paper; however, to appeal to a broader audience, it's good to give this level of detail in your repo description and README.
