### 2025-06-21 c97b3b6
Enhance README.md by expanding the description of ETLForge to include a detailed explanation of ETL processes, improving clarity and context for users.
