### 2025-07-02 00085b4
docs: add PyPI downloads badge to README
