### 2025-07-18 #8439 @kkartas
@crvernon the repository shows a shorter public development history because the package evolved through several stages before its current form on GitHub. The core functionality was initially developed across various projects I worked on, where I created these ETL components to solve specific data processing challenges.
As these components proved valuable across multiple use cases, I decided to refine and consolidate them into a cohesive package. Before the current public GitHub repository, ETLForge existed as part of a private project with limited visibility. When preparing for open-source release, I reorganized the code, improved documentation and established a more structured architecture to ensure it would be useful to the broader community.
This explains why the public repository history appears relatively brief compared to the actual development timeline of the underlying code and concepts.
Please let me know if you need any additional information about the development process.
