### 2025-10-30 81b90f9
Refine paper.md to enhance the description of ETLForge's typical workflow, validation algorithms, and performance characteristics. Update limitations section for clarity on dataset size, nested structures, and statistical validation capabilities, ensuring comprehensive documentation for users.
