+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
   - name: Independent Researcher
     index: 1
-date: 18 July 2024
+date: 18 July 2025
 bibliography: paper.bib
 paper_type: software
 version: 1.0.3
@@ -48,10 +48,6 @@ The landscape of data generation and validation tools shows clear specialization
 | YAML/JSON schema support | Yes | No | Python/YAML | Python only | Python only |
 | Lightweight dependencies | Yes (6 core) | Yes (1 core) | No (20+ deps) | Yes (5 core) | Yes (0 core) |
 
-**Performance characteristics**: ETLForge generates 10,000 rows in ~6 seconds and validates 100,000 rows in ~2 seconds on standard hardware, making it suitable for CI/CD integration where Great Expectations may be too heavyweight for simple validation tasks.
-
-**Limitations compared to existing tools**: Unlike Great Expectations, ETLForge does not provide data profiling, drift detection, or advanced statistical validations. Unlike pandera, it lacks integration with type checkers and advanced pandas DataFrame validation. ETLForge prioritizes simplicity and schema consistency over advanced features.
-
 To our knowledge, no existing open-source project provides an integrated, schema-first workflow covering both generation and validation with a unified configuration format.
 
 ## Software description
@@ -63,34 +59,68 @@ ETLForge implements a dual-purpose architecture where a single YAML/JSON schema 
 - `DataValidator`: Validates CSV/Excel files against schema rules, returning detailed error reports
 - CLI interface: Enables command-line automation via Click [@Click2023]
 
-**Integration approach**: Rather than replacing existing tools, ETLForge complements ETL testing workflows by ensuring test data and validation rules remain synchronized. It integrates with pandas-based pipelines and exports to common formats (CSV, Excel via openpyxl).
+**Typical workflow**: ETLForge is designed to support the following pipeline:
 
-## Quality control
+1. **Schema definition**: Define data structure and constraints based on target system requirements
+2. **Test data generation**: Generate synthetic datasets for initial ETL pipeline development and unit testing
+3. **Pipeline validation**: Use the same schema to validate production or external data after ETL transformations
+4. **Quality assurance**: Identify discrepancies between expected schema constraints and actual data quality
 
-GitHub Actions run comprehensive checks on Python 3.9-3.11:
+This workflow demonstrates that while ETLForge generates synthetic test data, its primary value proposition is in the validation phase where real production data is checked against expected constraints. The generation capability primarily serves to create controlled test datasets for unit testing ETL transformations before production data becomes available.
 
-* **Unit tests** (`pytest`), achieving **77% line coverage** across 587 statements (`pytest --cov`)
-  - CLI module: 67% coverage (31/95 statements missing)  
-  - Generator module: 74% coverage (64/244 statements missing)
-  - Validator module: 86% coverage (33/241 statements missing)
-* **Static analysis** (`flake8`, `black --check`, `mypy`) with type checking
-* **Integration testing** via end-to-end example (`example.py`) validating complete workflows
-* **Performance benchmarks** tracking generation/validation speed regressions
+**Integration approach**: Rather than replacing existing tools, ETLForge complements ETL testing workflows by ensuring test data and validation rules remain synchronized. It integrates with pandas-based pipelines and exports to common formats (CSV, Excel via openpyxl). The framework targets tabular data structures, which represent the majority of ETL use cases in relational database and data warehouse environments.
 
-All checks complete in under 90 seconds on Ubuntu runners, supporting rapid development cycles.
+## Software methodology
 
-**Known limitations**: 
-- Large datasets (>1M rows) may require memory optimization
-- Complex nested data structures are not supported
-- Advanced statistical validations require integration with specialized tools
[…]
