### 2025-10-31 #9326 @Nelly-Barret
Hi @kkartas! 👋

Thank you for the prompt changes and the clarifications, it is much more clear now. Could you please re-generate the PDF of your paper with the command `@editorialbot generate pdf`?
