### 2025-10-31 kkartas/ETLForge#1 @Nelly-Barret
Can't run ETLForge
Dear @kkartas ,

I am trying to use ETLForge (MacBook Pro, 2019) but it seems that the `yaml` dependency is missing as show below:

```
❯ git clone https://github.com/kkartas/ETLForge.git
Cloning into 'ETLForge'...
remote: Enumerating objects: 345, done.
remote: Total 345 (delta 198), reused 248 (delta 102), pack-reused 0 (from 0)
Receiving objects: 100% (345/345), 402.30 KiB | 5.83 MiB/s, done.
Resolving deltas: 100% (198/198), done.

❯ cd ETLForge

❯ pip install etl-forge
Collecting etl-forge
  Downloading etl_forge-1.0.3-py3-none-any.whl.metadata (12 kB)
Requirement already satisfied: pandas>=1.3.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge) (2.2.2)
Requirement already satisfied: pyyaml>=5.4.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge) (6.0.2)
Collecting click>=8.0.0 (from etl-forge)
  Downloading click-8.3.0-py3-none-any.whl.metadata (2.6 kB)
Requirement already satisfied: openpyxl>=3.0.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge) (3.1.5)
Requirement already satisfied: numpy>=1.21.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge) (2.0.1)
Requirement already satisfied: psutil>=5.9.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge) (7.0.0)
Requirement already satisfied: et-xmlfile in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from openpyxl>=3.0.0->etl-forge) (1.1.0)
Requirement already satisfied: python-dateutil>=2.8.2 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from pandas>=1.3.0->etl-forge) (2.9.0.post0)
Requirement already satisfied: pytz>=2020.1 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from pandas>=1.3.0->etl-forge) (2024.1)
Requirement already satisfied: tzdata>=2022.7 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from pandas>=1.3.0->etl-forge) (2024.1)
Requirement already satisfied: six>=1.5 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from python-dateutil>=2.8.2->pandas>=1.3.0->etl-forge) (1.16.0)
Downloading etl_forge-1.0.3-py3-none-any.whl (17 kB)
Downloading click-8.3.0-py3-none-any.whl (107 kB)
Installing collected packages: click, etl-forge
Successfully installed click-8.3.0 etl-forge-1.0.3
[notice] A new release of pip is available: 25.1.1 -> 25.3
[notice] To update, run: pip install --upgrade pip

❯ etl-forge --version
etl-forge, version 1.0.3

❯ python example.py
Traceback (most recent call last):
  File "example.py", line 9, in <module>
    from etl_forge import DataGenerator, DataValidator
  File "/Users/nbarret/Documents/boulot/reviews/JOSS/ETLForge/etl_forge/__init__.py", line 6, in <module>
    from .generator import DataGenerator
  File "/Users/nbarret/Documents/boulot/reviews/JOSS/ETLForge/etl_forge/generator.py", line 6, in <module>
    import yaml
ModuleNotFoundError: No module named 'yaml'
```

I also tried to install ETLForge with more complete setup but it does not work for me:

```
❯ pip install "etl-forge[faker]"
Collecting etl-forge[faker]
  Using cached etl_forge-1.0.3-py3-none-any.whl.metadata (12 kB)
Requirement already satisfied: pandas>=1.3.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge[faker]) (2.2.2)
Requirement already satisfied: pyyaml>=5.4.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge[faker]) (6.0.2)
Requirement already satisfied: click>=8.0.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge[faker]) (8.3.0)
Requirement already satisfied: openpyxl>=3.0.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge[faker]) (3.1.5)
Requirement already satisfied: numpy>=1.21.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge[faker]) (2.0.1)
Requirement already satisfied: psutil>=5.9.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge[faker]) (7.0.0)
Requirement already satisfied: faker>=15.0.0 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from etl-forge[faker]) (27.4.0)
Requirement already satisfied: python-dateutil>=2.4 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from faker>=15.0.0->etl-forge[faker]) (2.9.0.post0)
Requirement already satisfied: et-xmlfile in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from openpyxl>=3.0.0->etl-forge[faker]) (1.1.0)
Requirement already satisfied: pytz>=2020.1 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from pandas>=1.3.0->etl-forge[faker]) (2024.1)
Requirement already satisfied: tzdata>=2022.7 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from pandas>=1.3.0->etl-forge[faker]) (2024.1)
Requirement already satisfied: six>=1.5 in /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages (from python-dateutil>=2.4->faker>=15.0.0->etl-forge[faker]) (1.16.0)
Using cached etl_forge-1.0.3-py3-none-any.whl (17 kB)
Installing collected packages: etl-forge
Successfully installed etl-forge-1.0.3
[notice] A new release of pip is available: 25.1.1 -> 25.3
[notice] To update, run: pip install --upgrade pip

❯ python example.py
Traceback (most recent call last):
  File "example.py", line 9, in <module>
    from etl_forge import DataGenerator, DataValidator
  File "/Users/nbarret/Documents/boulot/reviews/JOSS/ETLForge/etl_forge/__init__.py", line 6, in <module>
    from .generator import DataGenerator
  File "/Users/nbarret/Documents/boulot/reviews/JOSS/ETLForge/etl_forge/generator.py", line 6, in <module>
    import yaml
ModuleNotFoundError: No module named 'yaml'
```

Could you please tell me how I can install this missing `yaml` dependency? Thank you!
