### 2025-10-31 #9326 @Nelly-Barret
Hi again @kkartas! 👋

The paper looks better now, thank you!

Remaining points concerning the paper:
- [ ] I don't really understand the sentence "but lacks integration" in the sentence "The landscape of data generation and validation tools show clear specialization [...]".
- [ ] It may also be useful to describe/discuss a bit the table that summarizes existing approaches (in a few sentences).
- [ ] I would rename the section "Limitations compared to existing tools" into "Discussion"

The rest of the paper is good to me.

Concerning the software:
- [ ] I could install ETLForge but I can't run the `example.py` file because the `yaml` dependency is missing. See issue https://github.com/kkartas/ETLForge/issues/1
