### 2025-11-01 2057534
Update example.py and README.md with troubleshooting steps for ModuleNotFoundError related to 'yaml'. Add requirements.txt for core dependencies and development setup instructions.
