### 2025-11-01 0775a26
Refine paper.md to clarify the integration of data generation and validation in ETLForge, emphasizing its unique schema-driven approach. Update sections on performance characteristics and limitations to enhance user understanding of the tool's capabilities and trade-offs compared to existing solutions.
