+++ paper/paper.md
@@ -37,7 +37,7 @@ Existing libraries focus on either data generation (e.g., *Faker* [@Faker2024]) 
 
 ## State of the field
 
-The landscape of data generation and validation tools shows clear specialization but lacks integration:
+The landscape of data generation and validation tools shows clear specialization, with tools excelling at either generation or validation, but lacking integration between these two complementary tasks:
 
 | Capability | ETLForge | Faker | Great Expectations | pandera | Cerberus |
 |------------|----------|-------|-------------------|---------|----------|
@@ -48,11 +48,11 @@ The landscape of data generation and validation tools shows clear specialization
 | YAML/JSON schema support | Yes | No | Python/YAML | Python only | Python only |
 | Lightweight dependencies | Yes (6 core) | Yes (1 core) | No (20+ deps) | Yes (5 core) | Yes (0 core) |
 
-To our knowledge, no existing open-source project provides an integrated, schema-first workflow covering both generation and validation with a unified configuration format.
+This comparison highlights that while several mature tools exist for data validation (Great Expectations, pandera, Cerberus) and Faker provides excellent data generation capabilities, none integrate both functions under a single schema definition. ETLForge is the only tool in this comparison that supports schema-driven generation *and* validation using the same configuration file, eliminating the need to maintain parallel schema definitions. Great Expectations offers the most comprehensive validation features but requires substantially more dependencies, making it less suitable for lightweight CI/CD environments. Faker requires manual scripting to define generation patterns rather than declarative schemas. To our knowledge, no existing open-source project provides an integrated, schema-first workflow covering both generation and validation with a unified configuration format.
 
 ## Software description
 
-ETLForge implements a dual-purpose architecture where a single YAML/JSON schema drives both data generation and validation processes. The schema format supports common data types (integer, float, string, date, category), constraints (ranges, uniqueness, nullability), and realistic data generation via Faker integration.
+ETLForge implements a dual-purpose architecture where a single YAML/JSON schema drives both data generation and validation processes. The schema format supports common data types (integer, float, string, date, category), constraints (ranges, uniqueness, nullability) and realistic data generation via Faker integration.
 
 **Core components:**
 - `DataGenerator`: Creates synthetic datasets using pandas [@McKinney2010] and numpy [@Harris2020] for numerical operations
@@ -85,13 +85,13 @@ This workflow demonstrates that while ETLForge generates synthetic test data, it
 
 1. **Schema conformance**: Verifies all required columns exist and no unexpected columns are present
 2. **Type checking**: Validates each cell's data type matches schema specifications
-3. **Constraint validation**: Checks range constraints, uniqueness requirements, and categorical value memberships
+3. **Constraint validation**: Checks range constraints, uniqueness requirements and categorical value memberships
 4. **Null validation**: Ensures null values only appear in nullable fields
 5. **Error aggregation**: Collects all validation failures with row and column identifiers for detailed reporting
 
 The validation process short-circuits on structural errors (missing columns) but continues through all rows to provide comprehensive error reports. This design prioritizes complete feedback over early termination.
 
-**Quality control**: GitHub Actions run comprehensive checks on Python 3.9-3.11, including unit tests achieving 77% line coverage across 587 statements, static analysis via flake8 and mypy, and integration testing through end-to-end workflows. All checks complete in under 90 seconds on Ubuntu runners, supporting rapid development cycles.
+**Quality control**: GitHub Actions run comprehensive checks on Python 3.9-3.11, including unit tests achieving 77% line coverage across 587 statements, static analysis via flake8 and mypy and integration testing through end-to-end workflows. All checks complete in under 90 seconds on Ubuntu runners, supporting rapid development cycles.
 
 ## Performance characteristics
 
[…]
