### 2025-11-01 #9326 @kkartas
Hi @nbarret! 👋

Thank you for the thorough review and for testing the software! I've addressed all your points.

**Paper Updates:**

I've made the three changes you requested:

1. **Clarified "but lacks integration"** - Now reads: "The landscape of data generation and validation tools shows clear specialization, with tools excelling at either generation or validation, but lacking integration between these two complementary tasks"

2. **Added table discussion** - Added a paragraph after the table explaining how ETLForge compares to existing tools, highlighting that while mature validation tools exist (Great Expectations, pandera, Cerberus) and Faker provides generation, none integrate both under a single schema.

3. **Renamed to "Discussion"** - Changed the section title from "Limitations compared to existing tools" to "Discussion" and expanded the opening to better contextualize the trade-offs.

**Software Issue:**

I've addressed the YAML dependency issue and replied with solutions in #1. I've also updated the README with a troubleshooting section to help others avoid this problem.

All changes are pushed. Let me know if you need anything else!
