### 2025-11-12 631bc27
Enhance DataGenerator class in generator.py to support unique value generation with Faker templates. Implement logic to enforce uniqueness and handle fallback to random strings, improving data integrity and usability for ETL processes.
