### 2025-11-12 8373702
Refactor random string generation in DataGenerator class for improved readability. Split long lines into multiple lines to enhance code clarity while maintaining functionality.
