### 2025-11-12 dbc73d3
Refactor DataGenerator class to simplify unique value set initialization by removing type hint, enhancing code readability without altering functionality.
