### 2025-11-12 fe68075
Update version to 1.0.4 in CITATION.cff, __init__.py, paper.md, and test_integration.py to reflect the latest release. Ensure consistency across documentation and codebase.
