+++ etl_forge/__init__.py
@@ -7,6 +7,6 @@ from .generator import DataGenerator
 from .validator import DataValidator
 from .exceptions import ETLForgeError
 
-__version__ = "1.0.3"
+__version__ = "1.0.4"
 
 __all__ = ["DataGenerator", "DataValidator", "ETLForgeError"]
