### 2025-11-12 490163f
Refactor DataGenerator class to improve unique value generation logic, enhancing clarity and maintainability. Update related tests to ensure functionality remains intact.
