### 2025-11-13 f084649
Enhance CI workflow by adding XML coverage report generation and integrating Codecov for coverage uploads. This improves visibility into test coverage metrics and ensures better reporting in CI pipelines.
