+++ .github/workflows/ci.yml
@@ -51,7 +51,16 @@ jobs:
 
     - name: Test with pytest
       run: |
-        pytest tests/ --cov=etl_forge --cov-report=term-missing -v
+        pytest tests/ --cov=etl_forge --cov-report=term-missing --cov-report=xml -v
+
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v5
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        file: ./coverage.xml
+        flags: unittests
+        name: codecov-umbrella
+        fail_ci_if_error: false
 
   build:
     needs: test
