### 2025-11-13 aa3c842
Add Codecov badge to README for enhanced test coverage visibility
