### 2025-11-13 30088ea
New tests for increased coverage.
