+++ tests/test_generator_faker.py
@@ -1,0 +1,176 @@
+"""
+Tests for Faker integration and unique constraint enforcement.
+"""
+
+import pytest
+from etl_forge.generator import DataGenerator
+from etl_forge.exceptions import ETLForgeError
+
+
+class TestFakerIntegration:
+    """Test cases for Faker template integration."""
+
+    def test_unique_faker_email(self):
+        """Test unique email generation with Faker."""
+        schema = {
+            "fields": [
+                {
+                    "name": "email",
+                    "type": "string",
+                    "unique": True,
+                    "faker_template": "email",
+                }
+            ]
+        }
+        generator = DataGenerator(schema)
+        df = generator.generate_data(100)
+
+        # All emails should be unique
+        assert df["email"].nunique() == 100
+        assert len(df) == 100
+
+    def test_unique_faker_name(self):
+        """Test unique name generation with Faker."""
+        schema = {
+            "fields": [
+                {
+                    "name": "name",
+                    "type": "string",
+                    "unique": True,
+                    "faker_template": "name",
+                }
+            ]
+        }
+        generator = DataGenerator(schema)
+        df = generator.generate_data(50)
+
+        assert df["name"].nunique() == 50
+
+    def test_non_unique_faker_values(self):
+        """Test non-unique Faker values (should allow duplicates)."""
+        schema = {
+            "fields": [
+                {
+                    "name": "color",
+                    "type": "string",
+                    "unique": False,
+                    "faker_template": "color_name",
+                }
+            ]
+        }
+        generator = DataGenerator(schema)
+        df = generator.generate_data(100)
+
+        # Should generate data, duplicates are allowed
+        assert len(df) == 100
+
+    def test_faker_invalid_template_with_unique(self):
+        """Test that invalid Faker template falls back to random strings with unique constraint."""
+        schema = {
+            "fields": [
+                {
+                    "name": "field",
+                    "type": "string",
+                    "unique": True,
+                    "faker_template": "nonexistent_faker_method",
+                    "length": {"min": 10, "max": 15},
+                }
+            ]
+        }
+        generator = DataGenerator(schema)
+        df = generator.generate_data(50)
+
+        # Should fall back to random strings
+        assert len(df) == 50
+        assert df["field"].nunique() == 50
+        # Check string length
+        assert all(10 <= len(val) <= 15 for val in df["field"])
+
+    def test_faker_invalid_template_non_unique(self):
+        """Test that invalid Faker template falls back to random strings without unique constraint."""
+        schema = {
+            "fields": [
+                {
+                    "name": "field",
+                    "type": "string",
+                    "unique": False,
+                    "faker_template": "nonexistent_faker_method",
+                    "length": {"min": 8, "max": 12},
+                }
+            ]
+        }
+        generator = DataGenerator(schema)
+        df = generator.generate_data(30)
+
+        assert len(df) == 30
+        assert all(8 <= len(val) <= 12 for val in df["field"])
+
+    def test_unique_faker_insufficient_variety(self):
+        """Test error when Faker cannot generate enough unique values."""
+        schema = {
+            "fields": [
+                {
+                    "name": "bool_str",
+                    "type": "string",
+                    "unique": True,
+                    "faker_template": "boolean",  # Only generates True/False
+                }
+            ]
+        }
+        generator = DataGenerator(schema)
+
+        # Should fail - can't generate 100 unique boolean strings
+        with pytest.raises(
+            ETLForgeError,
+            match="Could not generate .* unique values .* using faker template",
+        ):
+            generator.generate_data(100)
+
+    def test_faker_without_faker_installed(self):
+        """Test Faker template when faker is not available."""
+        schema = {
+            "fields": [
+                {
[…]
