### 2025-11-13 620e720
Refactor test assertions in CLI coverage tests for improved readability by breaking long lines into multiple lines. Clean up unnecessary trailing whitespace in test_generator_faker.py.
