### 2025-12-01 2104e82
Enhance date column generation to support custom date formats and add corresponding unit tests for validation.
