### 2025-12-01 720f6bb
Fix date format configuration in date column generation test for consistency
