### 2025-12-01 d76c662
Remove unused ValidationError class and related error handling methods from DataValidator
