### 2025-12-01 ad79499
Clarify description of existing libraries and their schema management in the paper
