### 2025-12-02 cf6c972
Improve test for unique Faker values by simulating insufficient variety and adjusting parameters for clarity
