### 2025-12-02 1c17111
faker is needed for the test_generator_faker tests
