+++ pyproject.toml
@@ -61,3 +61,4 @@ license-files = []
 
 [tool.setuptools.dynamic]
 version = {attr = "etl_forge.__version__"} 
+
