### 2025-12-02 kkartas/ETLForge#9 @gwbischof
Paper Comment
I don't think this statement is accurate:

https://github.com/kkartas/ETLForge/blob/bedccabe1cb9a24e40b87df38295ccba90d86826/paper/paper.md?plain=1#L51

Specifically:
```
ETLForge is the only tool in this comparison that supports schema-driven generation *and* validation using the same configuration file, eliminating the need to maintain parallel schema definitions.
```

See the pandera docs: https://pandera.readthedocs.io/en/stable/data_synthesis_strategies.html#data-synthesis-strategies

With Pandera, you can do something like:
```
schema = pa.DataFrameSchema({"column1": pa.Column(str, coerce=True)})
validated_df = schema.validate(df)
schema.example(n=5)
```

I would recommend reviewing the rest of the paper for for statements similar to this one.

https://github.com/openjournals/joss-reviews/issues/9326
