### 2025-12-05 #9326 @gwbischof
Hi @kkartas,

A couple subjective thoughts that aren't needed for the review:

1. Your project focuses on tabular/dataframe data I would consider explaining that in the project description.

2. The `validate()` signature (as well as other functions) accepts both paths and DataFrames. For a library, I'd consider accepting only DataFrames and letting application code handle I/O. You would be able to remove the code for reading and writing to excel files, and you could remove the `openpyxl` dependency.
```
def validate(self, data_path: Union[str, Path, pd.DataFrame]) -> ValidationResult
```

3. Consider supporting an established schema standard like [Frictionless Table Schema](https://specs.frictionlessdata.io//table-schema/) or [JSON Schema](https://github.com/python-jsonschema/jsonschema).

@yuanqing-wang, It's not totally clear to me that this project/paper meets the requirements defined [here](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort) I'll leave that up to the editorial team to decide.

Review complete! 😄
