### 2025-12-12 bdf7b4f
Enhance documentation to specify that ETLForge validates ETL outputs for tabular data (pandas DataFrames) across pyproject.toml, README.md, and paper.md files.
