### 2025-12-12 6e81a77
Refactor data validation to accept pandas DataFrames instead of file paths in example.py, README.md, and cli.py. Update pyproject.toml to include openpyxl as an optional dependency for Excel support. Enhance tests to validate error handling for non-DataFrame inputs.
