### 2025-12-12 ba26db5
Add Frictionless and JSON Schema examples, and implement comprehensive tests for schema adapters
- Created `frictionless_schema.json` with employee record fields and constraints.
- Added `jsonschema_example.json` for employee records using JSON Schema format.
- Developed extensive tests in `test_schema_adapter.py` for detecting and converting schemas between ETLForge, Frictionless, and JSON Schema formats.
- Included tests for data generation and validation using both Frictionless and JSON Schema.
