+++ etl_forge/schema_adapter.py
@@ -1,0 +1,672 @@
+"""
+Schema adapter module for converting established schema standards to ETLForge format.
+
+This module provides adapters for:
+- Frictionless Table Schema (https://specs.frictionlessdata.io/table-schema/)
+- JSON Schema (https://json-schema.org/)
+
+These adapters allow ETLForge to consume schemas defined in widely-adopted
+standards, improving interoperability with other data tools.
+"""
+
+from typing import Dict, Any, List, Optional, Union
+from pathlib import Path
+import json
+import yaml
+from .exceptions import ETLForgeError
+
+
+class SchemaAdapter:
+    """
+    Base class for schema adapters.
+    
+    Schema adapters convert schemas from established standards
+    to ETLForge's internal format.
+    """
+    
+    @staticmethod
+    def detect_schema_type(schema: Dict[str, Any]) -> str:
+        """
+        Detect the type of schema based on its structure.
+        
+        Args:
+            schema: A dictionary containing the schema definition.
+            
+        Returns:
+            One of: 'etlforge', 'frictionless', 'jsonschema', or 'unknown'
+        """
+        if not isinstance(schema, dict):
+            return 'unknown'
+        
+        # ETLForge native format
+        if 'fields' in schema and isinstance(schema.get('fields'), list):
+            # Check if it looks like ETLForge format (has 'type' with our types)
+            fields = schema['fields']
+            if fields and isinstance(fields[0], dict):
+                first_field = fields[0]
+                if 'type' in first_field:
+                    etlforge_types = {'int', 'float', 'string', 'date', 'category'}
+                    if first_field['type'] in etlforge_types:
+                        return 'etlforge'
+                    # Frictionless uses different type names
+                    frictionless_types = {'integer', 'number', 'boolean', 'object', 
+                                          'array', 'datetime', 'time', 'year', 
+                                          'yearmonth', 'duration', 'geopoint', 'geojson'}
+                    if first_field['type'] in frictionless_types or first_field['type'] == 'string':
+                        # Could be either, check for frictionless-specific fields
+                        if any(key in first_field for key in ['constraints', 'format', 'rdfType']):
+                            return 'frictionless'
+                        # Check if it has ETLForge-specific fields
+                        if any(key in first_field for key in ['range', 'faker_template', 'null_rate', 'values']):
+                            return 'etlforge'
+                        return 'frictionless'  # Default to frictionless if unclear
+        
+        # JSON Schema format
+        if '$schema' in schema or 'properties' in schema:
+            return 'jsonschema'
+        
+        return 'unknown'
+    
+    @staticmethod
+    def load_and_convert(schema_path: Union[str, Path, dict]) -> Dict[str, Any]:
+        """
+        Load a schema from a file or dict and convert to ETLForge format.
+        
+        This method auto-detects the schema type and applies the appropriate
+        conversion.
+        
+        Args:
+            schema_path: The path to a schema file or a dictionary.
+            
+        Returns:
+            A schema dictionary in ETLForge format.
+            
+        Raises:
+            ETLForgeError: If the schema cannot be loaded or converted.
+        """
+        if isinstance(schema_path, dict):
+            schema = schema_path
+        else:
+            schema_path_obj = Path(schema_path)
+            if not schema_path_obj.exists():
+                raise ETLForgeError(f"Schema file not found at: {schema_path}")
+            
+            suffix = schema_path_obj.suffix.lower()
+            try:
+                with open(schema_path_obj, "r", encoding="utf-8") as file:
+                    if suffix in [".yaml", ".yml"]:
+                        schema = yaml.safe_load(file) or {}
[…]
