+++ etl_forge/__init__.py
@@ -6,7 +6,15 @@ outputs.
 from .generator import DataGenerator
 from .validator import DataValidator
 from .exceptions import ETLForgeError
+from .schema_adapter import SchemaAdapter, FrictionlessAdapter, JsonSchemaAdapter
 
 __version__ = "1.0.4"
 
-__all__ = ["DataGenerator", "DataValidator", "ETLForgeError"]
+__all__ = [
+    "DataGenerator",
+    "DataValidator",
+    "ETLForgeError",
+    "SchemaAdapter",
+    "FrictionlessAdapter",
+    "JsonSchemaAdapter",
+]

[…]
