### 2025-12-12 57e9632
Refactor code for improved readability and consistency
- Simplified error messages in cli.py for file not found exceptions.
- Removed unnecessary blank lines in generator.py and schema_adapter.py for cleaner code.
- Enhanced schema detection logic in schema_adapter.py for better clarity.
- Streamlined conversion methods in schema_adapter.py and added comments for better understanding.
- Updated tests in test_schema_adapter.py and test_validator.py to remove redundant blank lines and improve readability.
- Ensured consistent formatting and style across the codebase.
