+++ tests/test_schema_adapter.py
@@ -137,7 +137,7 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.convert(frictionless)
-        
+
         assert len(result["fields"]) == 4
         assert result["fields"][0]["type"] == "int"
         assert result["fields"][1]["type"] == "string"
@@ -169,13 +169,13 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.convert(frictionless)
-        
+
         id_field = result["fields"][0]
         assert id_field["nullable"] == False
         assert id_field["unique"] == True
         assert id_field["range"]["min"] == 1
         assert id_field["range"]["max"] == 1000
-        
+
         username_field = result["fields"][1]
         assert username_field["length"]["min"] == 3
         assert username_field["length"]["max"] == 20
@@ -194,7 +194,7 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.convert(frictionless)
-        
+
         assert result["fields"][0]["type"] == "category"
         assert result["fields"][0]["values"] == ["active", "inactive", "pending"]
 
@@ -206,7 +206,7 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.convert(frictionless)
-        
+
         assert result["fields"][0]["type"] == "category"
         assert result["fields"][0]["values"] == ["true", "false"]
 
@@ -218,7 +218,7 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.convert(frictionless)
-        
+
         assert result["fields"][0]["type"] == "date"
         assert result["fields"][0]["format"] == "%Y-%m-%d"
 
@@ -230,7 +230,7 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.convert(frictionless)
-        
+
         assert result["fields"][0]["type"] == "int"
 
     def test_convert_preserves_description(self):
@@ -245,7 +245,7 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.convert(frictionless)
-        
+
         assert result["fields"][0]["_description"] == "Primary identifier"
 
     def test_convert_rejects_unsupported_types(self):
@@ -303,13 +303,13 @@ class TestFrictionlessAdapter:
             ]
         }
         result = FrictionlessAdapter.to_frictionless(etl_schema)
-        
+
         assert result["fields"][0]["type"] == "integer"
         assert result["fields"][0]["constraints"]["required"] == True
         assert result["fields"][0]["constraints"]["unique"] == True
         assert result["fields"][0]["constraints"]["minimum"] == 1
         assert result["fields"][0]["constraints"]["maximum"] == 100
-        
+
         assert result["fields"][1]["constraints"]["enum"] == ["active", "inactive"]
 
 
@@ -328,7 +328,7 @@ class TestJsonSchemaAdapter:
             "required": ["id"],
         }
         result = JsonSchemaAdapter.convert(json_schema)
-        
+
         fields_by_name = {f["name"]: f for f in result["fields"]}
         assert fields_by_name["id"]["type"] == "int"
         assert fields_by_name["id"]["nullable"] == False
@@ -354,7 +354,7 @@ class TestJsonSchemaAdapter:
             },
         }
         result = JsonSchemaAdapter.convert(json_schema)
-        
+
         fields_by_name = {f["name"]: f for f in result["fields"]}
         assert fields_by_name["age"]["range"]["min"] == 0
         assert fields_by_name["age"]["range"]["max"] == 120
@@ -374,7 +374,7 @@ class TestJsonSchemaAdapter:
             },
         }
         result = JsonSchemaAdapter.convert(json_schema)
[…]
