+++ etl_forge/schema_adapter.py
@@ -19,68 +19,93 @@ from .exceptions import ETLForgeError
 class SchemaAdapter:
     """
     Base class for schema adapters.
-    
+
     Schema adapters convert schemas from established standards
     to ETLForge's internal format.
     """
-    
+
     @staticmethod
     def detect_schema_type(schema: Dict[str, Any]) -> str:
         """
         Detect the type of schema based on its structure.
-        
+
         Args:
             schema: A dictionary containing the schema definition.
-            
+
         Returns:
             One of: 'etlforge', 'frictionless', 'jsonschema', or 'unknown'
         """
         if not isinstance(schema, dict):
-            return 'unknown'
-        
+            return "unknown"
+
         # ETLForge native format
-        if 'fields' in schema and isinstance(schema.get('fields'), list):
+        if "fields" in schema and isinstance(schema.get("fields"), list):
             # Check if it looks like ETLForge format (has 'type' with our types)
-            fields = schema['fields']
+            fields = schema["fields"]
             if fields and isinstance(fields[0], dict):
                 first_field = fields[0]
-                if 'type' in first_field:
-                    etlforge_types = {'int', 'float', 'string', 'date', 'category'}
-                    if first_field['type'] in etlforge_types:
-                        return 'etlforge'
+                if "type" in first_field:
+                    etlforge_types = {"int", "float", "string", "date", "category"}
+                    if first_field["type"] in etlforge_types:
+                        return "etlforge"
                     # Frictionless uses different type names
-                    frictionless_types = {'integer', 'number', 'boolean', 'object', 
-                                          'array', 'datetime', 'time', 'year', 
-                                          'yearmonth', 'duration', 'geopoint', 'geojson'}
-                    if first_field['type'] in frictionless_types or first_field['type'] == 'string':
+                    frictionless_types = {
+                        "integer",
+                        "number",
+                        "boolean",
+                        "object",
+                        "array",
+                        "datetime",
+                        "time",
+                        "year",
+                        "yearmonth",
+                        "duration",
+                        "geopoint",
+                        "geojson",
+                    }
+                    if (
+                        first_field["type"] in frictionless_types
+                        or first_field["type"] == "string"
+                    ):
                         # Could be either, check for frictionless-specific fields
-                        if any(key in first_field for key in ['constraints', 'format', 'rdfType']):
-                            return 'frictionless'
+                        if any(
+                            key in first_field
+                            for key in ["constraints", "format", "rdfType"]
+                        ):
+                            return "frictionless"
                         # Check if it has ETLForge-specific fields
-                        if any(key in first_field for key in ['range', 'faker_template', 'null_rate', 'values']):
-                            return 'etlforge'
-                        return 'frictionless'  # Default to frictionless if unclear
-        
+                        if any(
+                            key in first_field
+                            for key in [
+                                "range",
+                                "faker_template",
+                                "null_rate",
+                                "values",
+                            ]
+                        ):
+                            return "etlforge"
+                        return "frictionless"  # Default to frictionless if unclear
+
         # JSON Schema format
-        if '$schema' in schema or 'properties' in schema:
-            return 'jsonschema'
[…]
