### 2025-12-12 9dad562
Bump version to 1.1.0 in CITATION.cff, __init__.py, paper.md, and update integration tests for CLI version check
