+++ CITATION.cff
@@ -2,8 +2,8 @@ cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 type: software
 title: "ETLForge: A Python Framework for Synthetic Test Data Generation and ETL Pipeline Validation"
-version: 1.0.4
-date-released: 2025-06-15
+version: 1.1.0
+date-released: 2025-12-12
 url: "https://github.com/kkartas/ETLForge"
 repository-code: "https://github.com/kkartas/ETLForge"
 license: MIT
