### 2025-12-18 #9326 @yuanqing-wang
Could @gwbischof kindly review the changes? Thank you!
