### 2025-12-19 #9326 @kkartas
Ηελλο @gwbischof and thank you for the detailed and thoughtful feedback. I appreciate you taking the time to provide constructive criticism that will improve the quality of this work.

### On the Statement of Need

You are correct that the current framing is inaccurate. Pandera does support both validation and data synthesis from the same schema through its `schema.example()` interface. I will revise the statement of need to focus on accurate differentiators:

- ETLForge provides a **CLI interface** for both generation and validation, enabling non-Python automation in CI/CD pipelines
- The schema is **file-based (YAML/JSON)** rather than embedded in Python code, making it shareable across teams and tools
- **Lightweight dependency footprint** (6 core dependencies vs 20+ for Great Expectations)

I will conduct additional research to ensure these differentiators are accurate and not already addressed by other tools.

### On Schema Standards

Your point about JSON Schema being designed for nested data is well taken. Since ETLForge specifically targets **tabular data** (CSV/Excel), Frictionless Table Schema is indeed the more appropriate standard to consider.

I have two questions for clarification:

1. **Native adoption vs. adapter approach**: You suggest that if I support Table Schema, I should use it as the native format rather than converting to a custom schema. This would mean replacing the current ETLForge schema syntax entirely with Frictionless Table Schema syntax. Is this your recommendation? My concern is that Frictionless Table Schema doesn't natively support features like `faker_template` for realistic data generation. Would extending Table Schema with custom properties (as the spec allows) be acceptable?

2. **Scope of support**: If I adopt Frictionless Table Schema as the native format, should the project also aim to be interoperable with other Frictionless tools (like `frictionless-py`), or is alignment with the schema specification sufficient?

### On the Implementation

I acknowledge that the current adapter implementation added significant complexity. Rather than proceeding with that approach, I will:

1. Remove the JSON Schema support (it doesn't fit the tabular data focus)
2. Evaluate whether to adopt Frictionless Table Schema as the native format or maintain the custom format with clear justification

I want to make an informed decision here rather than rushing to implement. Could you share your perspective on whether extending Frictionless Table Schema with custom properties (for features like Faker integration) would be acceptable, or if a cleaner separation is preferred?

Thank you again for guiding this work toward higher quality.
