### 2026-01-19 #9326 @gwbischof
Hi @kkartas,

I think you have to address the comments about the Statement of Need accuracy.

@yuanqing-wang is support for a schema standard also a requirement?
