### 2026-02-09 2c30702
Update publication date in paper.md and refine summary and discussion sections for clarity and consistency
