+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
   - name: Independent Researcher
     index: 1
-date: 3 December 2025
+date: 9 Februrary 2026
 bibliography: paper.bib
 paper_type: software
 version: 1.1.0
@@ -23,47 +23,33 @@ version: 1.1.0
 
 ## Summary
 
-ETLForge is a lightweight Python package (Python ≥ 3.9) that generates realistic synthetic test data *and* validates ETL outputs for **tabular data** (pandas DataFrames) using the **same declarative schema**. One YAML/JSON file describes field types, ranges, uniqueness, nullability and optional Faker templates [@Faker2024]. The schema is consumed by two high-level components: `DataGenerator` (creates CSV/Excel datasets) and `DataValidator` (checks pipeline outputs, returning row-level error reports). A Click-based CLI [@Click2023] mirrors the library API, enabling automation within CI/CD workflows [@Fowler2013]. ETLForge therefore removes duplicated specifications and closes gaps where data-quality regressions can slip through testing cycles.
+ETLForge is a Python package (Python >= 3.9) for tabular ETL testing workflows. It uses one external YAML/JSON schema to drive both synthetic data generation and validation for pandas DataFrames. The schema can describe field types, ranges, uniqueness, nullability and optional Faker templates [@Faker2024]. ETLForge provides `DataGenerator` (creates tabular test datasets) and `DataValidator` (checks datasets and returns row-level error reports). A Click-based CLI [@Click2023] and a Python API expose the same core workflow, enabling local development and CI/CD automation [@Fowler2013].
 
 ## Statement of need
 
-Extract-Transform-Load (ETL) processes are critical for data-driven organizations, but testing these pipelines remains challenging [@Kimball2013; @Kleppmann2017]. Current testing approaches suffer from a fundamental disconnect: synthetic test data generation and output validation require separate tool chains with independent schema definitions. This creates several documented problems:
+Extract-Transform-Load (ETL) processes are critical for data-driven organizations, but testing these pipelines remains challenging [@Kimball2013; @Kleppmann2017]. Teams typically need both representative synthetic inputs and deterministic validation checks. Keeping those artifacts aligned over time can be labor-intensive and can introduce drift between test setup and quality checks [@Redman2016; @Dasu2003; @Loshin2010].
 
-1. **Schema drift**: When generation schemas diverge from validation rules, tests may pass while production data fails validation [@Redman2016].
-2. **Maintenance overhead**: Duplicate schema definitions require synchronized updates, increasing development time and error potential [@Dasu2003].
-3. **Testing gaps**: Inconsistent test data may not exercise edge cases that production validation catches, leading to false confidence [@Loshin2010].
-
-Most existing libraries focus primarily on either data generation (e.g., *Faker* [@Faker2024]) or validation (e.g., *Great Expectations* [@GreatExpectations2023]). While some tools like *pandera* [@Pandera2023] support both validation and data synthesis from the same schema (for example, via example-generation APIs), they are primarily designed around programmatic, Python-centric workflows rather than a declarative, file-based schema shared between a generator and a validator. In many teams this still leads to effectively **parallel specifications** for test data and validation rules, increasing the risk of drift and missed bugs. ETLForge unifies both stages under a single source of truth, reducing maintenance effort and improving test robustness. Its small dependency footprint (six runtime packages) fits comfortably inside continuous-integration pipelines.
[…]
