### 2026-02-09 #9326 @kkartas
Hi @gwbischof  , thank you for the follow-up.
I have updated the paper to correct the Statement of Need and removed the inaccurate “fundamental disconnect” claim.
The revised text now explicitly acknowledges existing tools (including pandera, Faker, Great Expectations, and Cerberus) and narrows ETLForge’s contribution to what is implemented: a configuration-first, tabular workflow where one external schema is used by both generation and validation, available via CLI and Python API.
I also removed high-risk comparison wording and aligned technical descriptions with the current implementation (including schema-adapter scope and validation behavior).
