### 2026-02-27 #9326 @yuanqing-wang
@kkartas could you double check your paper script to ensure that PDF generation is possible? thanks!
