### 2026-02-27 #9326 @kkartas
Hello @yuanqing-wang
I fixed the script and generated a pdf successfully.
