+++ CITATION.cff
@@ -2,8 +2,8 @@ cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 type: software
 title: "ETLForge: A Python Framework for Synthetic Test Data Generation and ETL Pipeline Validation"
-version: 1.1.0
-date-released: 2025-12-12
+version: 1.1.1
+date-released: 2026-02-27
 url: "https://github.com/kkartas/ETLForge"
 repository-code: "https://github.com/kkartas/ETLForge"
 license: MIT
@@ -36,7 +36,7 @@ preferred-citation:
       email: "mail@kkartas.gr"
       orcid: "https://orcid.org/0009-0001-6477-4676"
   journal: "Journal of Open Source Software"
-  year: 2025
+  year: 2026
   volume: TBD
   issue: TBD
   start: TBD
