### 2026-03-14 #9326 @crvernon
Editor note:  I have checked off the last box in the review checklist for @gwbischof per the response confirming a completed review [here](https://github.com/openjournals/joss-reviews/issues/9326#issuecomment-3618396714).
