### 2026-03-14 #9326 @kkartas
Hello @crvernon

I've changed the title of the archive to match the title of my paper.
The final version after all the reviews is v1.1.1 as mentioned [here](https://github.com/openjournals/joss-reviews/issues/9326#issuecomment-3972476563).
I've tried to change it, but I'm not allowed to.
