### 2026-03-14 1513f02
Update bibliography entries in paper.bib and paper.md for accuracy and consistency
