+++ paper/paper.bib
@@ -40,8 +40,7 @@
   author={{Faker Contributors}},
   year={2024},
   url={https://faker.readthedocs.io/},
-  version={24.4.0},
-  doi={10.5281/zenodo.3823398}
+  version={24.4.0}
 }
 
 @misc{Cerberus2024,
@@ -53,12 +52,12 @@
 }
 
 @misc{Pydantic2024,
-  title={Pydantic: Data validation using Python type hints},
-  author={{Pydantic Contributors}},
-  year={2024},
-  url={https://docs.pydantic.dev/},
+  title={pydantic/pydantic: v2.5.0 2023-11-13},
+  author={Colvin, Samuel and Montague, David and Garcia Badaracco, Adrian and Ramezani, Hasan and Jolibois, Eric and Trylesinski, Marcelo and Dorsey, Terrence and Runkle, Sydney and others},
+  year={2023},
+  url={https://github.com/pydantic/pydantic},
   version={2.5.0},
-  doi={10.5281/zenodo.12609353}
+  doi={10.5281/zenodo.10119276}
 }
 
 @article{McKinney2010,
@@ -104,12 +103,12 @@
 }
 
 @misc{GreatExpectations2023,
-  title={{Great Expectations: Always know what to expect from your data}},
-  author={{The Great Expectations Team}},
+  title={{Great Expectations}},
+  author={Gong, Abe and Campbell, James and {Great Expectations}},
   year={2023},
   url={https://greatexpectations.io/},
   version={0.18.0},
-  doi={10.5281/zenodo.14976223}
+  doi={10.5281/zenodo.10056226}
 }
 
 @software{pyyaml,
@@ -117,8 +116,7 @@
   title = "{PyYAML}",
   url = {https://pyyaml.org/},
   version = {6.0.1},
-  year = {2023},
-  doi={10.5281/zenodo.3873738}
+  year = {2023}
 }
 
 @misc{Click2023,
@@ -126,17 +124,16 @@
   author={{The Pallets Projects}},
   year={2023},
   url={https://click.palletsprojects.com/},
-  version={8.1.3},
-  doi={10.5281/zenodo.5123732}
+  version={8.1.3}
 }
 
-@misc{Pandera2023,
-  title={{pandera: Data validation for scientists, engineers, and analysts}},
-  author={{The pandera development team}},
-  year={2023},
-  url={https://pandera.readthedocs.io/},
-  version={0.17.0},
-  doi={10.25080/gerudo-f2bc6f59-010}
+@article{Pandera2020,
+  title={{pandera: Statistical Data Validation of Pandas Dataframes}},
+  author={Bantilan, Niels},
+  year={2020},
+  journal={Proceedings of the Python in Science Conference},
+  pages={116--124},
+  doi={10.25080/Majora-342d178e-010}
 }
 
  
