+++ paper/paper.bib
@@ -24,7 +24,8 @@
   title={The Practitioner's Guide to Data Quality Improvement},
   author={Loshin, David},
   year={2010},
-  publisher={Morgan Kaufmann}
+  publisher={Morgan Kaufmann},
+  doi={10.1016/C2009-0-17212-4}
 }
 
 @article{Fowler2013,
