### 2026-03-24 #9326 @crvernon
@kkartas - I just need you to address the following before I accept this submission for publication:

In your paper:

- [x] LINES 49-53:  It looks like you were trying to create a bullet point list under "Core components:" but this did not render.  Please check this out to see if this is the expected behavior.
- [x] LINE 148:  The "p" in "python" should be capitalized.  You can do this in your bib file by placing any characters you wish to maintain formatting for in curly brackets.
- [x] LINE 151:  The "p" in "python" should be capitalized.
- [x] LINE 158: The "w" in "wiley" should be capitalized.

Let me know when these changes have been made and I'll accept after confirmation.  Thank you.
