### 2026-03-24 27e7420
Paper formating hotfixes
