+++ paper/paper.bib
@@ -2,7 +2,7 @@
   title={The Data Warehouse Toolkit: The Definitive Guide to Dimensional Modeling},
   author={Kimball, Ralph and Ross, Margy},
   year={2013},
-  publisher={John wiley \& sons},
+  publisher={John {Wiley} \& Sons},
   edition={3rd}
 }
 
@@ -37,7 +37,7 @@
 }
 
 @misc{Faker2024,
-  title={Faker: A Python package that generates fake data for you},
+  title={Faker: A {Python} package that generates fake data for you},
   author={{Faker Contributors}},
   year={2024},
   url={https://faker.readthedocs.io/},
@@ -45,7 +45,7 @@
 }
 
 @misc{Cerberus2024,
-  title={Cerberus: Lightweight, extensible data validation library for Python},
+  title={Cerberus: Lightweight, extensible data validation library for {Python}},
   author={{Cerberus Contributors}},
   year={2024},
   url={https://docs.python-cerberus.org/},
@@ -62,9 +62,9 @@
 }
 
 @article{McKinney2010,
-  title={Data structures for statistical computing in Python},
+  title={Data structures for statistical computing in {Python}},
   author={McKinney, Wes and others},
-  journal={Proceedings of the 9th Python in Science Conference},
+  journal={Proceedings of the 9th {Python} in Science Conference},
   volume={445},
   pages={51--56},
   year={2010},
@@ -132,7 +132,7 @@
   title={{pandera: Statistical Data Validation of Pandas Dataframes}},
   author={Bantilan, Niels},
   year={2020},
-  journal={Proceedings of the Python in Science Conference},
+  journal={Proceedings of the {Python} in Science Conference},
   pages={116--124},
   doi={10.25080/Majora-342d178e-010}
 }
